+++ Inventors.Xml/Content/ChoiceElement.cs
@@ -6,8 +6,6 @@ using System.Threading.Tasks;
 
 namespace Inventors.Xml.Content
 {
-    public record Choice(string Name, Element Type);
-
     public class ChoiceElement :
         Element
     {
