--- Inventors.Xml/Content/EnumValueDescriptor.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Inventors.Xml.Content
-{
-    public record EnumValueDescriptor(string EnumValue, string Name);
-}
